Remove stale core sync and history DTOs
