+++ docs/history-source-plugins.md
@@ -105,10 +105,11 @@ An accepted import has one acquisition and publication path:
 Cold imports, appends, rewrites, replacements, and no-ops all use the shared
 custom JSONL source-family path. There is no fallback to the old Store database,
 synthetic `NativePath` body, command-output snapshot, or local content pack.
-`ctx show` resolves the active Core/Tantivy identity back to exact records in the
-provider-owned JSONL file. If that source is unavailable or no longer matches
-the certified locator, show reports typed unavailability instead of returning
-an indexed preview as complete.
+`ctx show` reads complete policy-selected normalized records from the active
+Core/Tantivy generation and does not reopen the provider-owned JSONL file. If a
+later import or refresh cannot read or validate that acquisition source, the
+last verified Core generation remains available until a successful refresh
+publishes a replacement.
 
 After import, `--history-source` uses the canonical
 `provider_key/source_id` route identity:
